Use injectStyle for custom css feature, and move it to core/common/css.ts
