+++ src/ts/core/common/css.ts
@@ -1,0 +1,11 @@
+export const injectStyle = (css: string, tagId: string) => {
+    if (document.getElementById(tagId)) {
+        document.getElementById(tagId)!.innerHTML = css
+        return
+    }
+
+    const style = document.createElement('style')
+    style.id = tagId
+    style.innerHTML = css
+    document.getElementsByTagName('head')[0].appendChild(style)
+}
